+++ alembic/env.py
@@ -3,8 +3,8 @@ from logging.config import fileConfig
 from sqlalchemy import engine_from_config, pool
 
 from alembic import context
-from src.config import settings
-from src.db.base import Base
+from src.core.config import settings
+from src.core.db.base import Base
 
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
